requirements.txt
flask
flask-cors
python-docx
python-dotenv
requests
pyngrok
mistune>=3.0.0

# MCP Server dependencies
fastmcp>=2.13.0
mcp>=1.22.0
